feat: edit Baan mode and refactor
